backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy>=2.0.2
pandas>=2.2.1
cloudconvert==2.1.0
openai==1.68.0
requests==2.31.0
pycryptodome
